Extract figures from HTML output
The default behavior of the HTMLExporter is to embed figures in the
HTML. This has a few downsides for performance, but the biggest issue is
Astropy Librarian can't reference these figures to include in the search
front end (an alternative approach would be to prepare special
thumbnails for the homepage; which is something we should do regardless
eventually).

This adds the ExtractOutputPreprocessor to the HTMLExporter
configuration. It's hard-coded at the moment. I did try looking at how
this might be added through the existing trait-based CLI parsing and it
wasn't clear to me how to support a list-based trait.

This is based on the tutorial:
https://nbconvert.readthedocs.io/en/latest/nbconvert_library.html#Extracting-Figures-using-the-HTML-Exporter