Add --preprocessors convert CLI option
This CLI option makes it possible to set additional preprocessors the
the HTML exporter. The main use case is to enable the
ExtractOutputPreprocessor (removing the hard coding from the previous
commit).

I think that technically this can also be done via the trait
configuration mechanism already in the CLI, but it would need to be
adapted to handle "list" trait types.