Apply ruff code suggestions
This is a first pass to make the codebase consistent with the Ruff
ruleset. I've added some exceptions to the Ruff rules in pyproject.toml
for some refactoring projects that will are more substantial,
particularly migrating from os.path to pathlib and also simplifying the
NbcollectionConverter's constructor method. We can tackle those later
and retire those exceptions.