Enable mypy type checking
This adds mypy for checking type annotations. Mypy is configured to
allow untyped code, which is necessary for allowing a partial migration
to type annotations. The new themes code will use type annotations.

Mypy is run in GitHub Actions ci.yaml and has a tox environment
(typing).

Mypy picked up on the assignment to a variable called "e" outside an
exception block in converter.py; apparently it has a rule against that,
so I renamed that variable to "exceptions".

Mypy also struggled with the dynamic base class technique used in
logger.py, so I added a comment for mypy to bypass that module for now.