Add CLI options and configs for Astropy template
The Learn Astropy template needs context information for the GitHub
repository that's hosting the notebooks. This information lets us
compute URLs for viewing/downloading the notebook from GitHub and
running the notebook on Binder.

To make it easier to manager these configuration, we're introducing a
new NbcollectionConfig class, which is a dataclass. This config class is
populated in the get_converter factory for Nbcollection, and the class
is passed as a keyword argument.

In the future it'll be beneficial to add more configurations to this
class, which will make configuration easier to pass to Nbcollection.