Use the LearnAstropyHtmlExporter
Instead of nbconvert's HtmlExporter, we now use the
LearnAstropyHtmlExporter (a subclass) that's defined in
nbcollection.themes.learnastropy. This exporter adds a UI around the
notebook for navigating to Learn Astropy and for download the notebook
file, viewing the notebook on GitHub, and opening the notebook in
Binder.

Note that as part of the conversion process to HTML we're now passing
information to the HTML template via the resources dictionary. We get
the information from these URLs through the NbcollectionConfig.

I've added a test to test_cli.py that inspects the generated HTML and
validates that the links are as-expected.

One thing I'd like to improve is the NbcollectionConverter's constructor
where the paths to the notebook files are resolved. There's a lot of
different cases for how to determine the "root" directory that
corresponds to the configured root path of the notebooks relative to a
GitHub repository.